requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Arpeggio==2.0.2
av==14.1.0
Backtesting==0.3.3
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bokeh==3.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clockblocks==0.6.9.post1
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
djitellopy==2.5.0
dnspython==2.7.0
email_validator==2.2.0
expenvelope==0.7.2.post2
fastapi==0.115.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fluidsynth==0.2
fonttools==4.54.1
frozendict==2.4.6
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib==3.9.2
MIDIUtil==1.2.1
mido==1.3.3
multitasking==0.0.11
networkx==3.4.2
numpy==2.1.2
openai==1.63.2
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pydantic==2.10.6
pydantic_core==2.27.2
pymusicxml==0.5.6
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-Quartz==10.3.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-osc==1.9.3
python-rtmidi==1.5.8
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scamp==0.9.2.post3
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.45.3
ta==0.11.0
tenacity==9.0.0
tornado==6.4.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.1.2
xyzservices==2024.9.0
yelpapi==2.5.1
yfinance==0.2.46
